requirements.txt
aiogram~=3.3.0
aiohttp~=3.9.0
sqlalchemy~=2.0.25
yookassa>=3.7.0
python-dotenv~=1.0.0